Stats and gameplay change
Now the gameplay is a bit different : ennemies won't reappear after they
pass the bottom of the screen. This is the first towards executing the
shooter-defense gameplay.

Also added a stats module to store the different kind of informations
during a game, such as player deaths, ennemy kills and of course ennemy
wins.